Move helper functions to utils directory (#386)
* Create utils dir

* Apply black
