+++ src/spyglass/common/common_sensors.py
@@ -7,8 +7,8 @@ from .common_ephys import Raw
 from .common_interval import IntervalList  # noqa: F401
 from .common_nwbfile import Nwbfile
 from .common_session import Session  # noqa: F401
-from .dj_helper_fn import fetch_nwb
-from .nwb_helper_fn import get_data_interface, get_nwb_file
+from ..utils.dj_helper_fn import fetch_nwb
+from ..utils.nwb_helper_fn import get_data_interface, get_nwb_file
 
 schema = dj.schema("common_sensors")
