+++ src/spyglass/decoding/sorted_spikes.py
@@ -23,7 +23,7 @@ from replay_trajectory_classification.initial_conditions import UniformInitialCo
 from spyglass.common.common_interval import IntervalList
 from spyglass.common.common_nwbfile import AnalysisNwbfile
 from spyglass.common.common_position import IntervalPositionInfo
-from spyglass.common.dj_helper_fn import fetch_nwb
+from spyglass.utils.dj_helper_fn import fetch_nwb
 from spyglass.decoding.core import (
     convert_epoch_interval_name_to_position_interval_name,
     convert_valid_times_to_slice,
